+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2022/06/28
 ## Version 0.22
 ### Changes
-* Changed dataFormatVersion to 10 to reflect minor change to percentageUtilization field from str() to float()
+* Changed dataFormatVersion to 10 to reflect minor change to percentageUtilization and queryTime field from str() to float()
 
 # 2022/06/26
 ## Version 0.21